migrating more types and logic
